Add separate file for tests
